refactor: shared/ui 분리
